Prevent thundering herd on websocket reconnect (#10960)
On large installations sometimes a proxy server can restart
or a network blip can happen which can disconnect a large
number of websocket connections. When this happens
all of those clients will simultaneously try to reconnect
putting the server under immense pressure.

This is fondly termed as thundering herd. And the solution
is to add a random time interval to spread out the reconnects.

We observed this happening on community where a large number
of websocket reconnects causes the webapp to freeze due to
spikes in CPU and goroutines on the server side.

Changelog: mattermost