MM-57885: Do not mark channel as read on tab unload (#26811)
This was spurious code left from old times. It's not
needed now, and it was causing a race condition with the server
side where we were setting the status to offline on
websocket disconnect. So if this request reached the server
after the status was set to offline, the status would
reset to online.

https://mattermost.atlassian.net/browse/MM-57885

```release-note
Fixed a bug where status would incorrectly get stuck to online
after user closes the tab.
```

# Conflicts:
#	webapp/channels/src/components/logged_in/logged_in.test.tsx

Changelog: removed